Updated manager scripts
Each crop type needs its own manager script, not just each pathway (C3, C4, CCM)
