+++ Scripts/WheatManagerCCM.cs
@@ -2,11 +2,8 @@ using System;
 using ModelFramework;
 
 using DCAPST;
-using DCAPST.Canopy;
-using DCAPST.Environment;
 using DCAPST.Interfaces;
-using DCAPST.Utilities;
- 
+using DCAPST.Utilities; 
 
 public class Script 
 {      
@@ -32,68 +29,74 @@ public class Script
    public DCAPSTModel DM;
    
    public double LAITrigger = 0.5;
-   public double PsiFactor = 0.4; //psiFactor-Psi Reduction Factor
-   
+   public double PsiFactor = 1.0; // Psi reduction factor
 
    // The following event handler will be called once at the beginning of the simulation
    [EventHandler] public void OnInitialised()
    {
+      /* IMPORTANT - Do NOT change the order of these values */
+
       CP = Classic.SetUpCanopy(
-         CanopyType.C4, // canopyType
-         363,     // airCO2
-         0.7,     // curvatureFactor
-         0.047,   // diffusivitySolubilityRatio
-         210000,  // airO2
-         0.78,    // diffuseExtCoeff
-         0.8,     // diffuseExtCoeffNIR
-         0.036,   // diffuseReflectionCoeff
-         0.389,   // diffuseReflectionCoeffNIR
-         60,      // leafAngle
-         0.15,    // leafScatteringCoeff
-         0.8,     // leafScatteringCoeffNIR
-         0.15,    // leafWidth
-         1.3,     // slnRatioTop
-         14,      // minimumN
-         1.5,     // windspeed
-         1.5      // windSpeedExtinction
+         CanopyType.CCM, // Canopy type
+         370,     // CO2 partial pressure
+         0.7,     // Empirical curvature factor
+         0.047,   // Diffusivity-solubility ratio
+         210000,  // O2 partial pressure
+         0.78,    // PAR diffuse extinction coefficient
+         0.8,     // NIR diffuse extinction coefficient
+         0.036,   // PAR diffuse reflection coefficient
+         0.389,   // NIR diffuse reflection coefficient
+         60,      // Leaf angle
+         0.15,    // PAR leaf scattering coefficient
+         0.8,     // NIR leaf scattering coefficient
+         0.05,    // Leaf width
+         1.3,     // SLN ratio at canopy top
+         14,      // Minimum structural nitrogen
+         1.5,     // Wind speed
+         1.5      // Wind speed profile distribution coefficient
       );
 
       PP = Classic.SetUpPathway(
-         0,                   // jTMin
-         37.8649150880407,    // jTOpt
-         55,                  // jTMax
-         0.711229539802063,   // jC
-         1,                   // jBeta
-         0,                   // gTMin
-         42,                  // gTOpt
-         55,                  // gTMax
-         0.462820450976839,   // gC
-         1,                   // gBeta
-         1210,    // KcAt25
-         64200,   // KcFactor
-         292000,  // KoAt25
-         10500,   // KoFactor
-         5.51328906454566, // VcVoAt25
-         21265.4029552906, // VcVoFactor
-         75,    // KpAt25
-         36300, // KpFactor
-         78000, // VcFactor
+         0,                   // Electron transport minimum temperature
+         30.0,                // Electron transport optimum temperature
+         45.0,                // Electron transport maximum temperature
+         0.911017958600129,   // Electron transport scaling constant
+         1,                   // Electron transport Beta value
+
+         0,                   // Mesophyll conductance minimum temperature
+         29.2338417788683,    // Mesophyll conductance optimum temperature
+         45,                  // Mesophyll conductance maximum temperature
+         0.875790608584141,   // Mesophyll conductance scaling constant
+         1,                   // Mesophyll conductance Beta value
+
+         273.422964228666,    // Kc25 - Michaelis Menten constant of Rubisco carboxylation at 25 degrees C
+         93720,               // KcFactor
+
+         165824.064155384,    // Ko25 - Michaelis Menten constant of Rubisco oxygenation at 25 degrees C
+         33600,               // KoFactor
+
+         4.59217066521612,    // VcVo25 - Rubisco carboxylation to oxygenation at 25 degrees C
+         35713.19871277176,   // VcVoFactor
+
+         75,        // Kp25 - Michaelis Menten constant of PEPc activity at 25 degrees C
+         36300,     // KpFactor
+
+         65330, // VcFactor
          46390, // RdFactor
          57043.2677590512, // VpFactor
-         120,     // pepRegeneration
-         0.15,    // spectralCorrectionFactor
-         0.1,     // ps2ActivityFraction
-         0.003,   // bundleSheathConductance
-         0.465 * PsiFactor,   // maxRubiscoActivitySLNRatio
-         2.7 * PsiFactor,     // maxElectronTransportSLNRatio
-         0.0 * PsiFactor,     // respirationSLNRatio
-         1.55 * PsiFactor,    // maxPEPcActivitySLNRatio
-         0.0135 * PsiFactor,  // mesophyllCO2ConductanceSLNRatio
-         2,    // extraATPCost
-         0.4,  // mesophyllElectronTransportFraction
-         0.45  // intercellularToAirCO2Ratio
-      );
 
+         400.0,     // PEPc regeneration
+         0.15,    // Spectral correction factor
+         0.1,     // Photosystem II activity fraction
+         0.5,   // Bundle sheath CO2 conductance
+         1.1 * PsiFactor,        // Max Rubisco activity to SLN ratio
+         1.85 * PsiFactor,       // Max electron transport to SLN ratio
+         0.0 * PsiFactor,        // Respiration to SLN ratio
+         0.373684157583268 * PsiFactor,        // Max PEPc activity to SLN ratio
+         0.00412 * PsiFactor,   // Mesophyll CO2 conductance to SLN ratio
+         0.75,    // Extra ATP cost
+         0.7   // Intercellular CO2 to air CO2 ratio
+      );
 
       //Set the LAI trigger
       MyPaddock.Set("laiTrigger", LAITrigger);
@@ -125,7 +128,6 @@ public class Script
       // Model the photosynthesis
       DCAPSTModel DM = Classic.SetUpModel(CP, PP, DOY, latitude, maxT, minT, radn);
       DM.DailyRun(lai, SLN, SWAvailable, RootShootRatio);
-      //DM.DailyRun(0.5, 1.0, 3.5, 1.3, 100);
       
       // Outputs
       RootShoot = RootShootRatio;
